# src/FinancialPortfolioAPI.API/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Financial Portfolio API",
    "version": "1.0.0",
    "description": "SQLite or SQL Server test-server API with transactional settlement, soft deletion, audit history and committed changes. USD and fixed demo prices. See /docs/guide."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "ApiKey": []
    }
  ],
  "paths": {
    "/api/v1/portfolio": {
      "get": {
        "summary": "List and filter portfolios",
        "operationId": "getapivportfolio",
        "parameters": [
          {
            "name": "clientId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/PortfolioStatus"
            }
          },
          {
            "name": "riskProfile",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/RiskProfile"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Version, reference or state conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create an empty USD portfolio",
        "operationId": "postapivportfolio",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Success",
            "headers": {
              "Location": {
                "description": "Resource URL",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Portfolio"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Version, reference or state conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePortfolioRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolio/{id}": {
      "get": {
        "summary": "Read a portfolio without changing it",
        "operationId": "getapivportfolioid",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Portfolio"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Version, reference or state conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Replace portfolio metadata with a version check",
        "operationId": "putapivportfolioid",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Portfolio"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Version, reference or state conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePortfolioRequest"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Soft-delete an empty portfolio and retain history",
        "operationId": "deletePortfolio",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "expectedVersion",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Portfolio"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Version conflict or portfolio is not empty"
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/portfolio/{id}/recalculate": {
      "post": {
        "summary": "Refresh fixed demonstration prices",
        "operationId": "postapivportfolioidrecalculate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Portfolio"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Version, reference or state conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/transaction/portfolio/{portfolioId}": {
      "get": {
        "summary": "List portfolio transactions",
        "operationId": "getapivtransactionportfolioportfolioId",
        "parameters": [
          {
            "name": "portfolioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Version, reference or state conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create pending transaction; identical reference and payload returns original",
        "operationId": "postapivtransactionportfolioportfolioId",
        "parameters": [
          {
            "name": "portfolioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "headers": {
              "Location": {
                "description": "Resource URL",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Version, reference or state conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    },
    "schemas": {
      "RiskProfile": {
        "type": "string",
        "enum": [
          "Conservative",
          "Moderate",
          "Aggressive",
          "VeryAggressive"
        ]
      },
      "PortfolioStatus": {
        "type": "string",
        "enum": [
          "Active",
          "Inactive",
          "PendingReview",
          "Restricted",
          "Closed"
        ]
      },
      "TransactionType": {
        "type": "string",
        "enum": [
          "Buy",
          "Sell",
          "Dividend",
          "Deposit",
          "Withdrawal"
        ]
      },
      "TransactionStatus": {
        "type": "string",
        "enum": [
          "Pending",
          "Executed",
          "Settled",
          "Failed",
          "Cancelled"
        ]
      },
      "CreatePortfolioRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "clientId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "clientName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "riskProfile": {
            "$ref": "#/components/schemas/RiskProfile"
          }
        },
        "required": [
          "name",
          "clientId",
          "clientName"
        ]
      },
      "UpdatePortfolioRequest": {
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "riskProfile": {
            "$ref": "#/components/schemas/RiskProfile"
          },
          "status": {
            "$ref": "#/components/schemas/PortfolioStatus"
          }
        },
        "required": [
          "expectedVersion",
          "name",
          "riskProfile",
          "status"
        ]
      },
      "VersionRequest": {
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        },
        "required": [
          "expectedVersion"
        ]
      },
...
      "CreateTransactionRequest": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "referenceNumber": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "symbol": {
            "type": "string",
            "maxLength": 20,
            "nullable": true
          },
          "shares": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "maximum": 1000000000,
            "default": 0
          },
          "price": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "maximum": 1000000000,
            "default": 0
          },
          "amount": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "maximum": 1000000000000,
            "default": 0
          },
          "commission": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "maximum": 1000000,
            "default": 0
          },
          "notes": {
            "type": "string",
            "maxLength": 500,
            "default": ""
          }
        },
        "required": [
          "type",
          "referenceNumber"
        ]
      },
      "Holding": {
        "type": "object",
        "properties": {
          "symbol": {
            "type": "string"
          },
          "companyName": {
            "type": "string"
          },
          "assetClass": {
            "type": "string",
            "enum": [
              "Equity",
              "FixedIncome",
              "Alternative",
              "Cash",
              "Commodity",
              "RealEstate",
              "Cryptocurrency"
            ]
          },
          "shares": {
            "type": "number",
            "format": "double"
          },
          "currentPrice": {
            "type": "number",
            "format": "double"
          },
          "costBasis": {
            "type": "number",
            "format": "double"
          },
          "lastPriceUpdate": {
            "type": "string",
            "format": "date-time"
          },
          "marketValue": {
            "type": "number",
            "format": "double"
          },
          "gainLoss": {
            "type": "number",
            "format": "double"
          },
          "gainLossPercentage": {
            "type": "number",
            "format": "double"
          },
          "portfolioId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "symbol",
          "companyName",
          "assetClass",
          "shares",
          "currentPrice",
          "costBasis",
          "lastPriceUpdate",
          "marketValue",
          "gainLoss",
          "gainLossPercentage"
        ]
      },
      "Portfolio": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "clientId": {
            "type": "string"
          },
          "clientName": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "cashBalance": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/PortfolioStatus"
          },
          "riskProfile": {
            "$ref": "#/components/schemas/RiskProfile"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastUpdatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "holdings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Holding"
            }
          },
          "totalValue": {
            "type": "number",
            "format": "double"
          },
          "unrealizedGainLoss": {
            "type": "number",
            "format": "double"
          },
          "allocation": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "format": "double"
            }
          },
          "isDeleted": {
            "type": "boolean"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "clientId",
          "clientName",
          "currency",
          "cashBalance",
          "status",
          "riskProfile",
          "version",
          "createdAt",
          "lastUpdatedAt",
          "holdings",
          "totalValue",
          "unrealizedGainLoss",
          "allocation"
        ]
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "portfolioId": {
            "type": "string",
            "format": "uuid"
          },
          "symbol": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "shares": {
            "type": "number",
            "format": "double"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "commission": {
            "type": "number",
            "format": "double"
          },
          "netAmount": {
            "type": "number",
            "format": "double"
          },
          "referenceNumber": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/TransactionStatus"
          },
          "transactionDate": {
            "type": "string",
            "format": "date-time"
          },
          "settlementDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "id",
          "portfolioId",
          "symbol",
          "type",
          "shares",
          "price",
          "amount",
          "commission",
          "netAmount",
          "referenceNumber",
          "notes",
          "status",
          "transactionDate",
          "settlementDate",
          "version"
        ]
      },
...
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "title",
          "status"
        ]
      },
      "PortfolioPage": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Portfolio"
            }
          }
        },
        "required": [
          "total",
          "offset",
          "limit",
          "items"
        ]
      },
      "TransactionPage": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            }
          }
        },
        "required": [
          "total",
          "offset",
          "limit",
          "items"
        ]
      },
...
    }
  }
}